Fix a subtle bug in logging module.
1 character in the logging format was wrong,
and it was causing logs to silent fail!

Had to get into pdb to realize the mis-step.
Now fixed.

Did a little cleanup too.
